types: get rid of 16-bit int/uint

+++ adcom1/content.go
@@ -20,7 +20,7 @@ type Content struct {
 	//   integer
 	// Definition:
 	//   Episode number.
-	Episode uint16 `json:"episode,omitempty"`
+	Episode int `json:"episode,omitempty"`
 
 	// Attribute:
 	//   title
